+++ web/app/application-setup.js
@@ -1,3 +1,0 @@
-var nucleusReact = require('../../assets/index');
-
-nucleusReact.stores.singlePanelManager.registerGlobalEventHandler();